Move methods related to ili2db args to IliArgs classes
