+++ QgisModelBaker/libqgsprojectgen/db_factory/db_uri.py
@@ -20,14 +20,7 @@ from abc import ABC, abstractmethod
 
 
 class DbUri:
+
     @abstractmethod
     def get_uri_from_conf(self, configuration, su):
-        pass
-
-    @abstractmethod
-    def get_db_args_from_conf(self, configuration, hide_password=False):
-        pass
-
-    @abstractmethod
-    def get_specific_params_schema_import(self):
         pass
